Split LinuxGLSwapChainContext into separate classes for Wayland and X11
