+++ sources/Renderer/OpenGL/Platform/Linux/LinuxGLSwapChainContext.cpp
@@ -1,15 +1,18 @@
 /*
  * LinuxGLSwapChainContext.cpp
  *
- * Copyright (c) 2015 Lukas Hermanns. All rights reserved.
+ * Copyright (c) 2025 Lukas Hermanns. All rights reserved.
  * Licensed under the terms of the BSD 3-Clause license (see LICENSE.txt).
  */
 
-#include "LinuxGLSwapChainContext.h"
-#include "LinuxGLContextX11.h"
+#include "../GLSwapChainContext.h"
+#include "LinuxGLSwapChainContextX11.h"
+
+#if LLGL_LINUX_ENABLE_WAYLAND
+#include "LinuxGLSwapChainContextWayland.h"
+#endif
 
 #include "../../../../Core/CoreUtils.h"
-#include "../../../../Core/Exception.h"
 #include <LLGL/Platform/NativeHandle.h>
 
 
@@ -54,120 +57,6 @@ bool GLSwapChainContext::MakeCurrentUnchecked(GLSwapChainContext* context)
 #endif
 }
 
-
-/*
- * LinuxX11GLSwapChainContext class
- */
-
-LinuxX11GLSwapChainContext::LinuxX11GLSwapChainContext(LinuxGLContextX11& context, Surface& surface) :
-    GLSwapChainContext { context                 },
-    glc_               { static_cast<GLXContext>(context.GetGLXContext()) }
-{
-    /* Get native window handle */
-    NativeHandle nativeHandle = {};
-    if (surface.GetNativeHandle(&nativeHandle, sizeof(nativeHandle)))
-    {
-        dpy_ = nativeHandle.x11.display;
-        wnd_ = nativeHandle.x11.window;
-    }
-    else
-        LLGL_TRAP("failed to get X11 Display and Window from swap-chain surface");
 }
-
-bool LinuxX11GLSwapChainContext::HasDrawable() const
-{
-    return (wnd_ != 0);
-}
-
-bool LinuxX11GLSwapChainContext::SwapBuffers()
-{
-    glXSwapBuffers(dpy_, wnd_);
-    return true;
-}
-
-void LinuxX11GLSwapChainContext::Resize(const Extent2D& resolution)
-{
-    // dummy
-}
-
-bool LinuxX11GLSwapChainContext::MakeCurrentGLXContext(LinuxX11GLSwapChainContext *context)
-{
-    if (context)
-        return glXMakeCurrent(context->dpy_, context->wnd_, context->glc_);
-    else
-        return glXMakeCurrent(nullptr, 0, 0);
-}
-
-} // /namespace LLGL
-
-/*
- * LinuxWaylandGLSwapChainContext class
- */
-
-#if LLGL_LINUX_ENABLE_WAYLAND
-
-#include "LinuxGLContextWayland.h"
-
-#include <wayland-egl-core.h>
-
-namespace LLGL {
-
-LinuxWaylandGLSwapChainContext::LinuxWaylandGLSwapChainContext(LinuxGLContextWayland& context, Surface& surface) :
-    GLSwapChainContext { context                 },
-    glc_               { static_cast<EGLContext>(context.GetEGLContext()) }
-{
-    /* Get native window handle */
-    NativeHandle nativeHandle = {};
-    if (surface.GetNativeHandle(&nativeHandle, sizeof(nativeHandle)))
-    {
-        EGLDisplay display = eglGetCurrentDisplay();
-
-        EGLAttrib surfaceAttribs[] = {
-            EGL_RENDER_BUFFER, EGL_BACK_BUFFER,
-            EGL_GL_COLORSPACE, EGL_GL_COLORSPACE_LINEAR,
-            EGL_NONE
-        };
-
-        LLGL::Extent2D size = surface.GetContentSize();
-
-        wl_egl_window* win = wl_egl_window_create(nativeHandle.wayland.window, size.width, size.height);
-
-        EGLSurface surface = eglCreatePlatformWindowSurface(display, context.GetEGLConfig(), win, surfaceAttribs);
-
-        dpy_ = display;
-        wnd_ = surface;
-    }
-    else
-        LLGL_TRAP("failed to get X11 Display and Window from swap-chain surface");
-}
-
-bool LinuxWaylandGLSwapChainContext::HasDrawable() const
-{
-    return (wnd_ != 0);
-}
-
-bool LinuxWaylandGLSwapChainContext::SwapBuffers()
-{
-    eglSwapBuffers(dpy_, wnd_);
-    return true;
-}
-
-void LinuxWaylandGLSwapChainContext::Resize(const Extent2D& resolution)
-{
-    // dummy
-}
-
-bool LinuxWaylandGLSwapChainContext::MakeCurrentEGLContext(LinuxWaylandGLSwapChainContext *context)
-{
-    if (context)
-        return eglMakeCurrent(context->dpy_, context->wnd_, context->wnd_, context->glc_);
-    else
-        return eglMakeCurrent(nullptr, nullptr, nullptr, nullptr);
-}
-
-} // /namespace LLGL
-
-#endif
-
 
 // ================================================================================
